# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "Callback": {
        "additionalProperties": false,
        "properties": {
          "binding": {
            "maxLength": 100,
            "minLength": 40,
            "title": "Binding",
            "type": "string"
          },
          "code": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "state": {
            "maxLength": 100,
            "minLength": 40,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "binding",
          "code"
        ],
        "title": "Callback",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionInput": {
        "additionalProperties": false,
        "properties": {
          "csrf": {
            "maxLength": 100,
            "minLength": 40,
            "title": "Csrf",
            "type": "string"
          },
          "session": {
            "maxLength": 100,
            "minLength": 40,
            "title": "Session",
            "type": "string"
          }
        },
        "required": [
          "session",
          "csrf"
        ],
        "title": "SessionInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AuthenticatedBearer": {
        "description": "Verified provider API token plus server-owned session in OIDC mode; private local credential only in explicit development mode.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "F01 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/access": {
      "post": {
        "operationId": "access_v1_auth_access_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Access V1 Auth Access Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Access",
        "tags": [
          "Identity"
        ]
      }
    },
    "/v1/auth/callback": {
      "post": {
        "operationId": "callback_v1_auth_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Callback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Callback V1 Auth Callback Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Callback",
        "tags": [
          "Identity"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Logout V1 Auth Logout Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Identity"
        ]
      }
    },
    "/v1/auth/start": {
      "post": {
        "operationId": "start_v1_auth_start_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Start V1 Auth Start Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Start",
        "tags": [
          "Identity"
        ]
      }
    },
...
  }
}